% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{dataframe}
\alias{dataframe}
\title{Generic function for class 'reg' or 'reg_y'}
\usage{
dataframe(x, ...)
}
\arguments{
\item{x}{A reg or reg_y object}

\item{\dots}{additional arguments}
}
\value{
A data frame of regression  result
}
\description{
Generic function to extracte concentrated results of regression models.
}
\examples{
data(diabetes)
head(diabetes)

reg_coxph<-reg(data = diabetes, x = c(1,4, 6),
y = 5, time = 2, factor = c(1, 3, 4), model = 'coxph')

dataframe(reg_coxph,save=FALSE)
}
\seealso{
\code{\link{dataframe.reg}},   \code{\link{reg}}
}
