% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.reg.R
\name{dataframe.reg}
\alias{dataframe.reg}
\title{Retrieve a data.frame from regression models}
\usage{
\method{dataframe}{reg}(x, save = FALSE, file = NULL, sep = ",",
  row.names = FALSE, ...)
}
\arguments{
\item{x}{A reg or reg_y object}

\item{save}{A logical, whether to save the concentrated result}

\item{file}{A character string naming a file, see \code{\link{write.table}} for more details, the default filepath is current working directory and the filename is the current time.}

\item{sep, row.names, \dots}{See \code{\link{write.table}} for more details}
}
\description{
Return a concentrated result of regression models.
}
\examples{
reg_glm<-reg(data = diabetes, x=c(1:4),y = 5,
factor = c(1, 3, 4), model = 'glm')

dataframe(reg_glm)
# dataframe(reg_glm, save = TRUE)
}
\seealso{
\code{\link{reg}}
}
