% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detail.reg.R
\name{detail.reg}
\alias{detail.reg}
\title{Retrieve detail results of univariate regression models}
\usage{
\method{detail}{reg}(x)
}
\arguments{
\item{x}{A reg object}
}
\description{
Retrieve detail results of univariate regression models
}
\examples{
data(diabetes)
head(diabetes)

reg_glm<-reg(data = diabetes, y = 5, factor = c(1, 3, 4), model = 'glm')
detail(reg_glm)
}
\seealso{
\code{\link{reg}}
}
