% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{display}
\alias{display}
\title{Generic function to display a data.frame or `reg` object}
\usage{
display(x, ...)
}
\arguments{
\item{x}{an object to display, a data.frame or a  `reg` class}

\item{\dots}{additional arguments}
}
\description{
\code{display} is a generic function to display summary information of varibles in a data.frame or summary result of `reg` class
}
\seealso{
\code{\link{display.data.frame}}, \code{\link{display.reg}}
}
