% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.reg.R
\name{display.reg}
\alias{display.reg}
\title{summary the result of univariate regression}
\usage{
\method{display}{reg}(x, cuts = c(0.001, 0.01, 0.05, 0.1, 1),
  all_var = TRUE, ...)
}
\arguments{
\item{x}{A reg object}

\item{cuts}{A vector of significance values.}

\item{all_var}{A logical, whether to display all variables at specific p value cut-off, if \code{FALSE}, only display the first 10 variables}

\item{\dots}{additional arguments}
}
\value{
Summary information of significant varibles.
}
\description{
return summary information of univariate regression.
}
\examples{
reg_glm<-reg(data = diabetes, y = 5, factor = c(1, 3, 4), model = 'glm')
display(reg_glm)
display(reg_glm, all_var = FALSE)
}
\seealso{
\code{\link{dataframe.reg}},
  \code{\link{display}},  \code{\link{display.reg}}
}
