% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuickReg.R
\name{QuickReg}
\alias{QuickReg}
\title{QuickReg}
\usage{
QuickReg(data, iv.vars, iv.vars.names, dv.vars, dv.vars.names, specifications,
  fixed.effects, fixed.effects.names, fixed.effects.specifications, robust.se,
  cluster, cluster.names, table.title, out.name, dynamic.out.name, html.only,
  silent, save.fits, demeaning.acceleration, stargazer.digits,
  stargazer.font.size, stargazer.style, stargazer.omit.stat, stargazer.type)
}
\arguments{
\item{data}{Data frame in which all model variables are located.}

\item{iv.vars}{Vector of independent variable names in dataset (e.g. c("gdppc", "pop"))}

\item{iv.vars.names}{(Optional) Vector of desired independent variable names in table output (e.g. c("GDP per capita", "Population")). Defaults to values in "iv.vars" if none provided.}

\item{dv.vars}{Vector of dependent variable in dataset (e.g. c("democracy", "war"))}

\item{dv.vars.names}{(Optional) Vector of desired dependent variable names in table output (e.g. c("Democracy (Boix-Rosato-Miller 2012)", "War (with at least 1000 battle deaths)")). Defaults to values in "dv.vars" if none provided.}

\item{specifications}{(Optional) List of desired regression specifications (selections of independent variables). The list of regression specifications are applied to all dependent variables. E.g. list(c(1), c(1,2), c(2))).}

\item{fixed.effects}{(Optional) Vector of desired fixed effect variable names in dataset (e.g. c("region", "year"))}

\item{fixed.effects.names}{(Optional) Vector of desired fixed effects labels in table output (e.g. c("Region FE", "Year FE")). Defaults to values in "fixed.effects" if none provided.}

\item{fixed.effects.specifications}{(Optional) List of desired fixed effect specifications (selections of independent variables). These specifications are applied in sequence from the first to last model. If the number of specifications is less than the number of models, all fixed effects are applied in the remaining columns by default. If none provided, defaults to all fixed effects in all models.}

\item{robust.se}{(Optional) If TRUE, returns robust standard errors calculated using a sandwich estimator from the "sandwich" package. Defaults to FALSE (i.e. normal standard errors).}

\item{cluster}{(Optional) Name of variable in dataset by which cluster-robust standard errors should be computed using the cluster.vcov command of the multiwayvcov package.}

\item{cluster.names}{(Optional) Desired name or label of clustering variable to be reported in table output (e.g. "Country" yields a note on the bottom of the table reading "Country-Clustered Standard Errors in Parenthesis"). If cluster specified but no "cluster.names" provided, "Cluster-Robust Standard Errors in Parenthesis" is reported.}

\item{table.title}{(Optional) Specifies the title of the table with regression output. Defaults to "QuickReg" plus the date and time of creation in parenthesis.}

\item{out.name}{(Optional) Specifies the output file name. Defaults to "QuickReg.html".}

\item{dynamic.out.name}{(Optional) If TRUE, adds date and time of creation in brackets between the out.name and the file extension (e.g. QuickReg (2017-04-05-14-01-27).html)}

\item{html.only}{(Optional) If TRUE, no latex output produced (only HTML table). Defaults to FALSE.}

\item{silent}{(Optional) If TRUE, no messages are returned by the function. Defaults to FALSE.}

\item{save.fits}{(Optional) If TRUE, saves fitted lm objects in a list by the name "QuickReg.fits" adding an integer if an object by this name already exists. Defaults to FALSE.}

\item{demeaning.acceleration}{(Optional) If TRUE, attempts to speed up regression by the method of alternating projections. In particular, it utilizes the "demeanlist" function of the "lfe" package to create a matrix of all covariates demeaned by all fixed effects, and then fits the different regression specifications on this demeaned matrix. Time saved is increasing in the number of fixed effects, specifications and observations, and this method is slower when all these are low. If there are thousands of fixed effects and many specifications, time saved is potentially quite large. Note: Overrides fixed.effects.specifications, always including all variables specified in fixed.effects, and does not supply R-squared or other model statistics. Defaults to FALSE.}

\item{...}{Various options passed to the stargazer function. See ?stargazer.}
}
\description{
The QuickReg package provides an easy interface for linear regression in R. This includes the option to request robust and clustered standard errors, automatic labeling, and an easy way to specify multiple regression specifications simulatenously, and a compact html or latex output (relying on the widely used "stargazer" package). It also includes several functionalities to speed up computation, including a special implementation of the method of alternating projections that may reduce calculation time by more than 50 percent for analysis with a large number of fixed effects, and whose performance gain is increasing in the number of specifications passed to the function simultaneously. Written by Sondre U. Solstad (ssolstad@princeton.edu).
}
\examples{
Please see: github.com/sondreus/QuickReg

}
\keyword{LS}
\keyword{OLS}
\keyword{QuickReg}
\keyword{cluster}
\keyword{lm}
\keyword{reg}
\keyword{regression}
\keyword{robust}
\keyword{robust.se}

