% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_y.R
\name{reg_y}
\alias{reg_y}
\title{Build regression models with more than one dependent varibale}
\usage{
reg_y(data = NULL, x = NULL, y = NULL, cov = NULL, factor = NULL,
  model = NULL, time = NULL, cov_show = FALSE, detail_show = FALSE)
}
\arguments{
\item{data}{A data.frame}

\item{x}{Integer column indices or names of the variables to be included in univariate analysis, the default columns are all the variables except `y` and `time` and `cov`.}

\item{y}{Integer column indices or name of dependent variable}

\item{cov}{Integer column indices or name of covariate variables}

\item{factor}{Integer column indices or names of variables to be treated as factor}

\item{model}{regression model, see \code{\link{lm}}, \code{\link{glm}}, \code{\link[survival]{coxph}} for more details}

\item{time}{Integer column indices  or name of survival time, used in cox regression, see \code{\link[survival]{coxph}} for more details}

\item{cov_show}{A logical, whether to create covariates result, default FALSE}

\item{detail_show}{A logical, whether to create each regression result, default FALSE}
}
\value{
The return result is a list including two componets, the first part is a detailed anaysis result, the second part is a concentrated result in a  data.frame
}
\description{
Build general linear model, generalized linear model, cox regression model,etc.
}
\examples{
reg_glm<-reg_y(data = diabetes, x = c(1:4, 6), y =c("diabetes","C2rs9332739","CFBrs641153") , factor = c(1, 3, 4), model = 'glm')
##  subset result like a list
reg_glm$dataframe
}
