% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detail.R
\name{detail}
\alias{detail}
\title{Generic function for class 'reg' or 'reg_y'}
\usage{
detail(x)
}
\arguments{
\item{x}{A reg or reg_y object}
}
\description{
Generic function to extracte every regression model result.
}
\examples{
data(diabetes)
head(diabetes)
reg_coxph<-reg(data = diabetes, x = c(1,4, 6), y = 5, time = 2, factor = c(1, 4), model = 'coxph')
detail(reg_coxph)
}
\seealso{
\code{\link{detail.reg}}, \code{\link{reg}}
}
