% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R, R/display.data.frame.R
\name{display}
\alias{display}
\alias{display}
\title{Generic function to display a data.frame or `reg` object}
\usage{
display(data = NULL, col = NULL, normtest = NULL, useNA = "ifany",
  discrete_limit = 10, exclude_discrete = TRUE, ...)

display(data = NULL, col = NULL, normtest = NULL, useNA = "ifany",
  discrete_limit = 10, exclude_discrete = TRUE, ...)
}
\arguments{
\item{data}{A data.frame}

\item{col}{Column indices or names of the variables in the dataset to display, the default columns are all the variables}

\item{normtest}{A character indicating test of normality, the default method is \code{\link{shapiro.test}} when sample size no more than 5000, otherwise \code{\link[nortest]{lillie.test}}{Kolmogorov-Smirnov} is used, see package \strong{nortest} for more methods.Use 'shapiro.test', 'lillie.test', 'ad.test', etc to specify methods.}

\item{useNA}{Whether to include NA values in the table, see \code{\link{table}} for more details}

\item{discrete_limit}{A numeric defining the minimal of unique value to display the variable as count and frequency}

\item{exclude_discrete}{Logical, whether to exclude discrete variables with more unique values specified by discrete_limit}

\item{\dots}{additional arguments}

\item{x}{an object to display, a data.frame or a  `reg` class}

\item{\dots}{additional arguments}
}
\description{
\code{display} is a generic function to display summary information of varibles in a data.frame or summary result of `reg` class

Dispaly count, frequency or mean, standard deviation and test of normality, etc.
}
\examples{
data(diabetes)
head(diabetes)
display(diabetes, 1:2)
display(diabetes, 1:11, normtest = "lillie.test")
}
\seealso{
\code{\link{display.data.frame}}, \code{\link{display.reg}}

\code{\link{display.reg}},  \code{\link{display}}
}
